requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
numpy==1.26.4
requests==2.31.0
skyfield==1.49
python-dotenv
